+++ src/client/client.js
@@ -117,7 +117,9 @@ Client.initialize = function (options) {
     );
   }
 
-  promise = getGatewayConfiguration(authData, options.sessionId).then(function (configuration) {
+  promise = getGatewayConfiguration(authData, options.sessionId).then(function (
+    configuration
+  ) {
     if (options.debug) {
       configuration.isDebug = true;
     }